Add partial support for enums
Partially implemented as described in the example from
http://hackage.haskell.org/package/QuickCheck-2.10.1/docs/Test-QuickCheck-Arbitrary.html#v:shrink.

Specifically, for variants with fields this implementation first shrinks
to unit variants, then directly goes to recursive shrinking, omitting
shrinking to subterms in case of recursive data structure - this needs
proper handling afterwards. For unit variants though it yields an empty
iterator - as expected.